update codes in views.py to validate ip addresses before saving it to the database and running the tool

+++ CREMEapplication/views.py
@@ -9,6 +9,9 @@ from .serializers import ProgressDataSerializer
 from rest_framework import viewsets
 
 from .tasks import execute_toolchain
+from django.contrib import messages
+import os
+
 
 # Create your views here.
 
@@ -44,6 +47,21 @@ def create_progress_data_if_not_exist():
     progress_datas = ProgressData.objects.all()
     if not progress_datas:
         ProgressData.objects.create()
+
+
+def validate_ips(ips):
+    errors = []
+    all_valid = True
+    for ip in ips:
+        if ' ' in ip:
+            all_valid = False
+            errors.append("({0}) is not a valid IP address".format(ip))
+        else:
+            HOST_UP = True if os.system("ping -c 1 " + ip) is 0 else False
+            if not HOST_UP:
+                all_valid = False
+                errors.append("Cannot connect with ({0})".format(ip))
+    return all_valid, errors
 
 
 def dashboard(request):
@@ -106,6 +124,77 @@ def new_testbed_information(request):
     num_of_malicious_client = testbed.number_of_malicious_client
 
     if request.method == "POST":
+        dict_machines = dict()
+        dict_machines['Controller'] = ControllerForm(request.POST, prefix='c')
+        dict_machines['Data Logger Server'] = DataLoggerServerForm(request.POST, prefix='dls')
+        dict_machines['Target Server'] = TargetServerForm(request.POST, prefix='ts')
+        dict_machines['Benign Server'] = BenignServerForm(request.POST, prefix='bs')
+        for i in range(num_of_vulnerable_client):
+            dict_machines['Vulnerable Client {0}'.format(i + 1)] = VulnerableClientForm(request.POST,
+                                                                                        prefix='vc{0}'.format(i + 1))
+        for i in range(num_of_non_vulnerable_client):
+            dict_machines['Non-Vulnerable Client {0}'.format(i + 1)] =\
+                NonVulnerableClientForm(request.POST, prefix='nvc{0}'.format(i + 1))
+        dict_machines['Attacker Server'] = AttackerServerForm(request.POST, prefix='as')
+        dict_machines['Malicious Client'] = MaliciousClientForm(request.POST, prefix='mc')
+
+        forms = []
+        ips = []
+
+        form_c = ControllerForm(request.POST, prefix='c')
+        if form_c.is_valid():
+            form = form_c.save(commit=False)
+            forms.append(form)
+
+        form_dls = DataLoggerServerForm(request.POST, prefix='dls')
+        if form_dls.is_valid():
+            form = form_dls.save(commit=False)
+            forms.append(form)
+
+        form_ts = TargetServerForm(request.POST, prefix='ts')
+        if form_ts.is_valid():
+            form = form_ts.save(commit=False)
+            forms.append(form)
+
+        form_bs = BenignServerForm(request.POST, prefix='bs')
+        if form_bs.is_valid():
+            form = form_bs.save(commit=False)
+            forms.append(form)
+
+        for i in range(num_of_vulnerable_client):
+            form_vc = VulnerableClientForm(request.POST, prefix='vc{0}'.format(i+1))
+            if form_vc.is_valid():
+                form = form_vc.save(commit=False)
+                forms.append(form)
+
+        for i in range(num_of_non_vulnerable_client):
+            form_nvc = NonVulnerableClientForm(request.POST, prefix='nvc{0}'.format(i+1))
+            if form_nvc.is_valid():
+                form = form_nvc.save(commit=False)
+                forms.append(form)
+
+        form_as = AttackerServerForm(request.POST, prefix='as')
+        if form_as.is_valid():
+            attacker_server = form_as.save(commit=False)
+            attacker_server.number_of_new_bots = num_of_vulnerable_client
+            forms.append(attacker_server)
+
+        form_mc = MaliciousClientForm(request.POST, prefix='mc')
+        if form_mc.is_valid():
+            form = form_mc.save(commit=False)
+            forms.append(form)
+
+        # get ips to check validation
+        for form in forms:
+            ips.append(form.ip)
+
+        # validate ip addresses
+        all_valid, errors = validate_ips(ips)
+        if not all_valid:
+            for error in errors:
+                messages.error(request, error)
+            return render(request, 'testbed/new_testbed_information.html', {'dict_machines': dict_machines})
+
         # clear all existing objects
         Controller.objects.all().delete()
         DataLoggerServer.objects.all().delete()
@@ -116,41 +205,9 @@ def new_testbed_information(request):
         AttackerServer.objects.all().delete()
         MaliciousClient.objects.all().delete()
 
-        form_c = ControllerForm(request.POST, prefix='c')
-        if form_c.is_valid():
-            form_c.save()
-
-        form_dls = DataLoggerServerForm(request.POST, prefix='dls')
-        if form_dls.is_valid():
-            form_dls.save()
-
-        form_ts = TargetServerForm(request.POST, prefix='ts')
-        if form_ts.is_valid():
-            form_ts.save()
-
-        form_bs = BenignServerForm(request.POST, prefix='bs')
-        if form_bs.is_valid():
-            form_bs.save()
-
-        form_as = AttackerServerForm(request.POST, prefix='as')
-        if form_as.is_valid():
-            attacker_server = form_as.save(commit=False)
-            attacker_server.number_of_new_bots = num_of_vulnerable_client
-            attacker_server.save()
-
-        form_mc = MaliciousClientForm(request.POST, prefix='mc')
-        if form_mc.is_valid():
-            form_mc.save()
-
-        for i in range(num_of_vulnerable_client):
-            form_vc = VulnerableClientForm(request.POST, prefix='vc{0}'.format(i+1))
-            if form_vc.is_valid():
-                form_vc.save()
-
-        for i in range(num_of_non_vulnerable_client):
-            form_nvc = NonVulnerableClientForm(request.POST, prefix='nvc{0}'.format(i+1))
-            if form_nvc.is_valid():
-                form_nvc.save()
+        # save forms if valid
+        for form in forms:
+            form.save()
 
         create_progress_data_if_not_exist()
